♻️ GH-83 Separate Config dataclass from validation_rule module
Group Document-archetype types (Plan, Config, SessionState)
away from Rule-archetype types so downstream M7 #D9 sub-packages
(domain/documents/ vs domain/rules/) have a clean starting line.

Solution:
- New domain/config_document.py holding the Config dataclass
- Remove Config from validation_rule.py
- Update 8 internal import sites
- domain/__init__.py still re-exports Config (backward-compat)

Refs: GH-83 #D4
