+++ src/dev10x/config/loader.py
@@ -8,10 +8,11 @@ from typing import Any
 import msgpack
 import yaml
 
+from dev10x.domain.config_document import Config
 from dev10x.domain.config_loader import ConfigLoader
 from dev10x.domain.file_locks import atomic_write_bytes
 from dev10x.domain.friction_level import FrictionLevel
-from dev10x.domain.validation_rule import Compensation, Config, Rule
+from dev10x.domain.validation_rule import Compensation, Rule
 
 DEFAULT_TTL_SECONDS = 1800
