+++ src/dev10x/domain/__init__.py
@@ -1,3 +1,4 @@
+from dev10x.domain.config_document import Config
 from dev10x.domain.config_loader import ConfigLoader
 from dev10x.domain.git_context import GitContext
 from dev10x.domain.hook_input import HookAllow, HookInput, HookResult, HookRetry
@@ -6,7 +7,7 @@ from dev10x.domain.repository_ref import RepositoryRef
 from dev10x.domain.result import ErrorResult, Result, SuccessResult, err, ok
 from dev10x.domain.rule_engine import RuleEngine, RuleMatch
 from dev10x.domain.sql import SqlStatement, is_read_only_sql
-from dev10x.domain.validation_rule import Compensation, Config, Rule
+from dev10x.domain.validation_rule import Compensation, Rule
 
 __all__ = [
     "Compensation",
